Handle possible null status on ManagedKafka
Without this change, there's an intermittent test failure due to a
NullPointerException on the call to
`managedKafka.getStatus().getVersions().getStrimzi()` when the status
is null.

I've updated it to use the same way of getting the version as in the
`hasStrimziChanged` method used in the `if` condition on the line
above.