Remove TopologySpreadConstraint toggle; set DoNotSchedule
The property defaults to true, and there currently isn't a scenario
where we need to set it to false. Given that there's also no clear
consensus on what it should be called, I'm removing it. If it's needed
in future, someone else can decide what to call it. :)

To explain why we don't ever need to not set a
TopologySpreadConstraint at the moment:

- In the default case of a cluster with 3 zones, we'd always want to
set it so that the brokers are always evenly spread across the 3
zones.
- In the case of a single zone cluster, it evenly spreads the 3
brokers across the 1 available zone, which means that all 3 brokers
get scheduled to the same zone.

Note that this commit also changes the "whenUnsatisfiable" value from
"ScheduleAnyway" to "DoNotSchedule". This prevents 2 brokers from
being scheduled to the same zone in a 3+ zone cluster, if there's some
transient reason that the pods can't be spread across all zones when
they're being created (like if there's a node down in one zone, and
not enough capacity left among the other nodes in the zone). This is
important because it can't be re-balanced across zones easily after
initial scheduling, because storage volumes aren't portable across
zones (at least in the clouds that we work with so far).