Don't wait for total readiness to add routes to CR
The canary now has an init container that waits for the bootstrap DNS
record to exist before proceeding with its normal operation. That init
container can block initialization for up to 10 minutes by
default (configurable timeout).

Unfortunately, without this change, the operator waits for all
operands to be ready before adding the route mapping to the
ManagedKafka CR, so the fleet manager never gets them and therefore
doesn't have the mappings to be able to create the DNS CNAME records,
until the canary init times out and the Pod is eventually marked
ready.

This change just waits until there are at least 5 routes (we currently
expect exactly 5), and they have non-empty targets set, then adds them
to the route mapping in the ManagedKafka CR.