Refactor StrimziBundleManager unit tests.
The tests previously relied on testing business logic by calling methods exposed specifically for the unit test.  The code paths where
the watcher actually called the business logic were untested making for a somewhat weak test.

The test is now refactored to use a simple watcher that mimics the important actions that would be made by OLM.  This
allows the code to be tested under conditions that match life much more closely.

Signed-off-by: kwall <kwall@apache.org>