Configure the soft and hard limits at customer's provisioned capacity.
We don't believe customers will notice producer throttling in time to act, a safety factor of 2 there would have been a little over 1 min of throttling before the hard limit. Moving to a factor of 5 would give them 2m 30s or so to react.
The hard limit is what really protects the broker anyway, so we just simplify the documentation and monitoring story to the hard limit.