Use new prefix for broker plugins properties
In the broker plugins project, we're updating the config properties so
that they don't have the strimzi prefix.

This change updates the config set by this operator so that both the
old and new properties are set. Once the version of the operator using
the new properties is rolled out, we can then remove the old
properties from being set here again in another change.

Given that this is intended to be a temporary duplication, I tried to
keep it as simple as possible, so that it will be simple to remove
when we get to that stage also.