MGDSTRM-10036: Add bf2_org_kafkaInstanceProfileType to Kafka instance metrics
ZookeeperContainersDown and KafkaBrokerContainersDown alerts rely on this label which is set on 'kafka_brokers' metric emitted from Kafka. However, if kafka containers are not running, this metric dissappear. In this situation, the alerts never trigger which is counter to their purpose of alerting when containers are not running.

Instead of 'kafka_brokers', the alerts can use kafka instance metrics emitted from fleetshard so that it doesn't rely on kafka itself.