Delete Strimzi cluster operator leadership lease on deployment deletion
Why:
As part of enabling StrimziPodSets in the cluster operator we are also increasing
availability of the operator by running 2 replicas and using a k8s lease to elect
one of them as a leader, the other will be in standby in case it needs to take over
reconciliation. The lease is not able to be automatically cleaned up when we tear
down a cluster operator deployment, so we are making the Fleetshard Operator responsible
for deleting the lease in reaction to a deployment deletion.

Note that the lease is named ${deploymentName}-leadership-token by convention to
enable fleetshard to clean it.

To get the test working I noticed that the StrimziManager and StrimziManagerTest were
communicating with different instances of the mock server on different ports. The
StrimziManager is wired with a client before the `@BeforeEach` is invoked. The beforeEach
triggered a new mock server to be created and used in the tests.